Changed keyboard/keymap level definitions
