+++ keyboards/edinburgh41/keymaps/lalit/config.h
@@ -3,20 +3,12 @@
 
 #pragma once
 
-// bootmagic settings
-
-#define BOOTMAGIC_LITE_ROW 0
-#define BOOTMAGIC_LITE_COLUMN 0
-
 // Underglow LED settings
 
-#define RGBLED_NUM 10
 #define RGBLIGHT_SLEEP
 
 // Thumbstick settings
 
-#define ANALOG_JOYSTICK_X_AXIS_PIN B5
-#define ANALOG_JOYSTICK_Y_AXIS_PIN B6
 #define SCROLLING_LAYER 1
 #define TAPPING_LAYER 2
 
@@ -25,6 +17,18 @@
 #define THUMBSTICK_UP_TAP KC_UP
 #define THUMBSTICK_DOWN_TAP KC_DOWN
 
+#define CURSOR_SPEED 90
+#define SCROLL_SPEED 20
+#define SCROLL_DELAY_MS 70
+
+// Thumbstick settings
+
+#define SCROLLING_LAYER 1
+#define TAPPING_LAYER 2
+// #define THUMBSTICK_RIGHT_TAP KC_RIGHT
+// #define THUMBSTICK_LEFT_TAP KC_LEFT
+// #define THUMBSTICK_UP_TAP KC_UP
+// #define THUMBSTICK_DOWN_TAP KC_DOWN
 #define CURSOR_SPEED 1
 #define SCROLL_SPEED 0.1
