--- keyboards/edinburgh41/post_config.h
@@ -1,40 +1,0 @@
-// Copyright 2022 L. Mistry (@schwarzer-geiger)
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-
-// default settings
-
-#ifndef RGBLED_NUM
-    #define RGBLED_NUM 10
-#endif
-#ifndef RGBLIGHT_DEFAULT_SAT
-    #define RGBLIGHT_DEFAULT_SAT 0
-#endif
-
-
-// Thumbstick default settings
-
-#ifndef SCROLLING_LAYER
-    #define SCROLLING_LAYER 1
-#endif
-#ifndef TAPPING_LAYER
-    #define TAPPING_LAYER 2
-#endif
-#ifndef THUMBSTICK_RIGHT_TAP
-    #define THUMBSTICK_RIGHT_TAP KC_RIGHT
-#endif
-#ifndef THUMBSTICK_LEFT_TAP
-    #define THUMBSTICK_LEFT_TAP KC_LEFT
-#endif
-#ifndef THUMBSTICK_UP_TAP
-    #define THUMBSTICK_UP_TAP KC_UP
-#endif
-#ifndef THUMBSTICK_DOWN_TAP
-    #define THUMBSTICK_DOWN_TAP KC_DOWN
-#endif
-#ifndef CURSOR_SPEED
-    #define CURSOR_SPEED 1
-#endif
-#ifndef SCROLL_SPEED
-    #define SCROLL_SPEED 0.1
-#endif
